Gavin Tasks Sprint 8
Server is live on Lightsail but I keep getting errors regarding read-only files when run normally. When I runserver with admin perms (sudo) it throws errors that didn't exist without admin perms. Not sure what the problem is but it seems we won't be able to demonstrate on the live server, just with a localhost.